# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Laravel API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List users",
                "description": "Paginated list with search and sorting. Allowed sort fields: id, name_lastname, name_firstname, email, created_at, updated_at",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "parameters": [
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "description": "Search by name or email (partial match)",
                        "schema": {
                            "type": "string"
                        },
                        "example": "田中"
                    },
                    {
                        "$ref": "#/components/parameters/QueryPage"
                    },
                    {
                        "$ref": "#/components/parameters/QueryPerPage"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field. Prefix `-` for descending.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "name_lastname",
                                "-name_lastname",
                                "name_firstname",
                                "-name_firstname",
                                "email",
                                "-email",
                                "created_at",
                                "-created_at",
                                "updated_at",
                                "-updated_at"
                            ]
                        },
                        "example": "-created_at"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated user list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "description": "Create a new user account",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user",
                "description": "Get user by ID",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PathId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "Update user (partial update supported)",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PathId"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Permanently delete user",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PathId"
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/NoContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserStoreRequest": {
                "required": [
                    "name_lastname",
                    "name_firstname",
                    "name_kana_lastname",
                    "name_kana_firstname",
                    "email",
                    "password"
                ],
                "properties": {
                    "name_lastname": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "田中"
                    },
                    "name_firstname": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "太郎"
                    },
                    "name_kana_lastname": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "タナカ"
                    },
                    "name_kana_firstname": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "タロウ"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "tanaka@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "name_lastname": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "山田"
                    },
                    "name_firstname": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "花子"
                    },
                    "name_kana_lastname": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "ヤマダ"
                    },
                    "name_kana_firstname": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "ハナコ"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "yamada@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "User resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name_lastname": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "田中"
                    },
                    "name_firstname": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "太郎"
                    },
                    "name_kana_lastname": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "タナカ"
                    },
                    "name_kana_firstname": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "タロウ"
                    },
                    "name_full_name": {
                        "description": "Computed: lastname + firstname",
                        "type": "string",
                        "example": "田中 太郎"
                    },
                    "name_full_name_kana": {
                        "description": "Computed: kana lastname + firstname",
                        "type": "string",
                        "example": "タナカ タロウ"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "tanaka@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "Success": {
                "description": "Successful operation"
            },
            "Created": {
                "description": "Resource created"
            },
            "NoContent": {
                "description": "Deleted"
            },
            "NotFound": {
                "description": "Not found"
            },
            "Unauthorized": {
                "description": "Unauthenticated"
            },
            "Forbidden": {
                "description": "Forbidden"
            },
            "ValidationError": {
                "description": "Validation failed",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "type": "object",
                                    "example": {
                                        "email": [
                                            "The email has already been taken."
                                        ]
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "QueryPage": {
                "name": "page",
                "in": "query",
                "description": "Page number",
                "schema": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                }
            },
            "QueryPerPage": {
                "name": "per_page",
                "in": "query",
                "description": "Items per page",
                "schema": {
                    "type": "integer",
                    "default": 10,
                    "maximum": 100,
                    "minimum": 1
                }
            },
            "PathId": {
                "name": "id",
                "in": "path",
                "description": "Resource ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Cookie",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "User management endpoints"
        }
    ]
}